@@ -5,14 +5,15 @@
     <modelVersion>4.0.0</modelVersion>
 
     <properties>
-        <spring-boot.version>2.1.4.RELEASE</spring-boot.version>
-        <spring-cloud.version>Greenwich.RELEASE</spring-cloud.version>
+        <spring-boot.version>2.5.5</spring-boot.version>
+        <spring-cloud.version>2020.0.4</spring-cloud.version>
         <rest-assured.version>3.1.0</rest-assured.version>
         <hazelcast.version>3.11.2</hazelcast.version>
         <lombok.version>1.18.6</lombok.version>
         <swagger.version>2.9.2</swagger.version>
         <!-- plugins -->
-        <java.version>1.8</java.version>
+<!--        <java.version>1.8</java.version>-->
+<!--        <java.version>15</java.version>-->
         <maven-surefire-plugin.version>2.19.1</maven-surefire-plugin.version>
         <maven-compiler-plugin.version>3.7.0</maven-compiler-plugin.version>
         <thin.version>1.0.21.RELEASE</thin.version>
@@ -21,7 +22,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.4.RELEASE</version>
+        <version>2.5.5</version>
     </parent>
 
     <groupId>com.oopsmails.spring.cloud.microservicesoauth2</groupId>
@@ -37,7 +38,7 @@
         <module>spring-cloud-microservices-eurekaserver</module>
         <module>spring-cloud-microservices-organizationservice</module>
         <module>spring-cloud-microservices-springgateway</module>
-        <module>spring-cloud-microservices-zuulproxy</module>
+<!--        <module>spring-cloud-microservices-zuulproxy</module>-->
     </modules>
 
     <dependencyManagement>
